リファクタリング中
・エラーの解決
・カラーバーのリファクタリング
・mainBaseのリファクタリング（途中）
