+++ Assets/Image/宝箱オープンエフェクトアニメ/unity_asset/TreasureChestOpenEffect.unitypackage.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5a62b418d4bc5a143b86805f3bfef5bf
+guid: b710ba395f885f24ba83e31e560fe518
 DefaultImporter:
   externalObjects: {}
   userData:
