+++ Assets/Scenes/osoujiKanrikunn/ios_push_aojiru.unity.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d902a6927588da94896c9c1695995baf
+guid: d2cdf7b235a32de40a7ebd2b354e1b86
 DefaultImporter:
   externalObjects: {}
   userData:
