+++ Assets/Script/CleanPlaceData/CleanPlaceData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e069d78d8bae5224db8a225febc6f461
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
