+++ Assets/cleanPlaceData.json
@@ -1,1 +1,1 @@
-{"placeDataList":[{"place":"test","lastUpdateTime_forSave":{"year":2020,"month":1,"day":9,"hour":0,"minute":0,"second":0},"cleanInterval_forSave":{"year":0,"month":0,"day":3,"hour":0,"minute":0,"second":0},"leftDate":-12},{"place":"test","lastUpdateTime_forSave":{"year":2020,"month":1,"day":6,"hour":0,"minute":0,"second":0},"cleanInterval_forSave":{"year":0,"month":0,"day":6,"hour":0,"minute":0,"second":0},"leftDate":-12},{"place":"uniti","lastUpdateTime_forSave":{"year":2019,"month":11,"day":15,"hour":0,"minute":0,"second":0},"cleanInterval_forSave":{"year":0,"month":0,"day":60,"hour":0,"minute":0,"second":0},"leftDate":-10},{"place":"a","lastUpdateTime_forSave":{"year":2020,"month":1,"day":9,"hour":0,"minute":0,"second":0},"cleanInterval_forSave":{"year":0,"month":0,"day":6,"hour":0,"minute":0,"second":0},"leftDate":-9},{"place":"hh","lastUpdateTime_forSave":{"year":2020,"month":1,"day":9,"hour":0,"minute":0,"second":0},"cleanInterval_forSave":{"year":0,"month":0,"day":6,"hour":0,"minute":0,"second":0},"leftDate":-9},{"place":"リビング","lastUpdateTime_forSave":{"year":2019,"month":12,"day":20,"hour":0,"minute":0,"second":0},"cleanInterval_forSave":{"year":0,"month":0,"day":26,"hour":0,"minute":0,"second":0},"leftDate":-9},{"place":"iiiiiiiiiiii","lastUpdateTime_forSave":{"year":2020,"month":1,"day":24,"hour":0,"minute":0,"second":0},"cleanInterval_forSave":{"year":0,"month":0,"day":4,"hour":0,"minute":0,"second":0},"leftDate":4},{"place":"大吾","lastUpdateTime_forSave":{"year":2020,"month":1,"day":24,"hour":0,"minute":0,"second":0},"cleanInterval_forSave":{"year":0,"month":0,"day":10,"hour":0,"minute":0,"second":0},"leftDate":10},{"place":"最終兵器俺たち","lastUpdateTime_forSave":{"year":2020,"month":1,"day":24,"hour":0,"minute":0,"second":0},"cleanInterval_forSave":{"year":0,"month":0,"day":28,"hour":0,"minute":0,"second":0},"leftDate":28},{"place":"a","lastUpdateTime_forSave":{"year":2020,"month":1,"day":7,"hour":0,"minute":0,"second":0},"cleanInterval_forSave":{"year":0,"month":0,"day":180,"hour":0,"minute":0,"second":0},"leftDate":163},{"place":"青じ","lastUpdateTime_forSave":{"year":2019,"month":8,"day":27,"hour":0,"minute":0,"second":0},"cleanInterval_forSave":{"year":0,"month":0,"day":5400,"hour":0,"minute":0,"second":0},"leftDate":5250}],"pushTiming":5}
+{"placeDataList":[{"place":"hh","lastUpdateTime_forSave":{"year":2020,"month":1,"day":9,"hour":0,"minute":0,"second":0},"cleanInterval_forSave":{"year":0,"month":0,"day":6,"hour":0,"minute":0,"second":0},"leftDate":-22},{"place":"test","lastUpdateTime_forSave":{"year":2020,"month":2,"day":6,"hour":0,"minute":0,"second":0},"cleanInterval_forSave":{"year":0,"month":0,"day":10,"hour":0,"minute":0,"second":0},"leftDate":1},{"place":"t","lastUpdateTime_forSave":{"year":2020,"month":1,"day":29,"hour":0,"minute":0,"second":0},"cleanInterval_forSave":{"year":0,"month":0,"day":10,"hour":0,"minute":0,"second":0},"leftDate":2},{"place":"t","lastUpdateTime_forSave":{"year":2020,"month":1,"day":30,"hour":0,"minute":0,"second":0},"cleanInterval_forSave":{"year":0,"month":0,"day":10,"hour":0,"minute":0,"second":0},"leftDate":3},{"place":"t","lastUpdateTime_forSave":{"year":2020,"month":1,"day":20,"hour":0,"minute":0,"second":0},"cleanInterval_forSave":{"year":0,"month":0,"day":20,"hour":0,"minute":0,"second":0},"leftDate":3},{"place":"t","lastUpdateTime_forSave":{"year":2020,"month":1,"day":31,"hour":0,"minute":0,"second":0},"cleanInterval_forSave":{"year":0,"month":0,"day":10,"hour":0,"minute":0,"second":0},"leftDate":4},{"place":"test","lastUpdateTime_forSave":{"year":2020,"month":2,"day":1,"hour":0,"minute":0,"second":0},"cleanInterval_forSave":{"year":0,"month":0,"day":10,"hour":0,"minute":0,"second":0},"leftDate":5},{"place":"t","lastUpdateTime_forSave":{"year":2020,"month":2,"day":2,"hour":0,"minute":0,"second":0},"cleanInterval_forSave":{"year":0,"month":0,"day":10,"hour":0,"minute":0,"second":0},"leftDate":6},{"place":"test","lastUpdateTime_forSave":{"year":2020,"month":2,"day":5,"hour":0,"minute":0,"second":0},"cleanInterval_forSave":{"year":0,"month":0,"day":10,"hour":0,"minute":0,"second":0},"leftDate":9}],"pushTiming":5}
